+++ ch07/seq2seq.py
@@ -79,7 +79,10 @@ class Decoder:
         self.lstm.set_state(h)
 
         for _ in range(sample_size):
-            x = np.array(sample_id).reshape((1, 1))
+
+            # 2020.01.03 ishida
+            # x = np.array(sample_id).reshape((1, 1))
+            x = np.array(sample_id , dtype=np.int).reshape((1, 1))
             out = self.embed.forward(x)
             out = self.lstm.forward(out)
             score = self.affine.forward(out)
